fix(stella-plugin): split the panel contract out of a file that met the ceiling
Three CI failures, two causes, both mine.

`wire_contract.rs` reached 1514 lines and `manifest.rs` 1506, and a file that
crosses 1500 gets a split rather than a baseline entry. The panel channel's
tests move to `tests/panel_contract.rs` — the seam is the channel and not the
line count, since a panel is its own dispatch context with its own point,
grant and vocabulary. Every one of the 38 names in the old file survives in
one of the two new ones; nothing was dropped in the move. The name-drawability
rule moves to `panel.rs` beside the predicate it shares with `PanelText::new`,
which is where it belonged anyway and takes `manifest.rs` back under.

`PluginManifest::validate` is private, so two of my doc links tripped
`-D rustdoc::private-intra-doc-links` and failed both rustdoc jobs. They now
use the `[`text`](crate::PublicType)` form already used at `panel.rs`'s
`PanelGrant::validate` and `oracle.rs`.

`check-wire-schema` itself was green throughout — that job failed on its
`doc-warnings-schema` step, not on a stale artifact.

Refs #5203, #5210
